backend/temp-test/requirements.txt
# Core web framework
fastapi==0.115.2
uvicorn==0.32.0
gunicorn
# Reddit API client
asyncpraw==7.8.1

# OpenAI SDK (async client)
openai==1.51.0
httpx==0.27.2
# Environment variable loader
python-dotenv==1.0.1

# Data validation
pydantic==2.9.2
